rename student-api to studentApi
